requirements.txt
numpy
opencv-python
Pillow
scikit-learn
scipy
six
tqdm